Add ExampleExperimentSpace component that allows users to clone one of the example experiments
